refactor(logger): move logger to separate folder
